Working on test coverage
