+++ src/fedservice/entity/client/list.py
@@ -3,9 +3,13 @@ from typing import Optional
 from typing import Union
 from urllib.parse import urlencode
 
+from cryptojwt.jws.jws import factory
 from idpyoidc.client.configure import Configuration
 from idpyoidc.message.oauth2 import ResponseMessage
 
+from fedservice.entity import get_verified_trust_chains
+from fedservice.entity.function.trust_anchor import get_verified_endpoint
+from fedservice.entity.function.trust_anchor import get_verified_trust_anchor_statement
 from fedservice.entity.service import FederationService
 from fedservice.entity.utils import get_federation_entity
 from fedservice.message import ListResponse
@@ -44,12 +48,7 @@ class List(FederationService):
         :return: List of entity IDs
         """
         if not endpoint:
-            _federation_entity = get_federation_entity(self)
-            _collector = _federation_entity.function.trust_chain_collector
-            _ec = _collector.config_cache[entity_id]
-            endpoint = _ec["metadata"]["federation_entity"].get(self.endpoint_name)
-            if not endpoint:
-                return None
+            endpoint = get_verified_endpoint(self, entity_id, self.endpoint_name)
 
         qpart = {}
         for arg in ["entity_type", "trust_marked", "trust_mark_id", "intermediate"]:
